lambda_generate/requirements.txt
openai
chromadb
boto3
numpy
aiohttp
asyncio
pydantic
aiofiles
tiktoken
requests
python-dotenv
google-generativeai
chonkie
aws-lambda-powertools
autotiktokenizer